+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -2,15 +2,17 @@
 
 ## 1.1.0-beta.2 (Unreleased)
 
-- Added `TransferResponseDisposeOwnership` to `PipelineMessage` to enable returning an undisposed `PipelineResponse` from protocol methods.
-
 ### Features Added
 
 ### Breaking Changes
 
+- Change `HttpClientPipelineTransport.Shared` from a field to a property.
+
 ### Bugs Fixed
 
 ### Other Changes
+
+- Made `CancellationToken` parameter passed to `BinaryContent.WriteTo` optional.
 
 ## 1.1.0-beta.1 (2024-02-01)